unified_backend/requirements.txt
Flask
flask_cors
scikit-learn
nltk
sentence-transformers
PyPDF2
openai-whisper
soundfile
transformers
torch
sentencepiece
